fix: Game: heavily refactored input(); made game-input TU
- input() renamed to "receive_agent_input".
- proces_intendedmove() renamed to "process_agent_input".
- getInput() (in "global"'s TU) renamed to "getKeypressDownInput".
- Moved check_input_{ansi,vim,wasd}() to new "game-input" TU and to the
namespace "Game::Input".
- Moved "keyboard value codes" enums and enum "IntendedMoveFlag" (with
the datatype "intendedmove_t") to new "game-input" TU and to the
namespace "Game::Input".
- check_input_{ansi,vim,wasd}(), receive_agent_input() and
process_agent_input() now modify a intendedmove_t struct via its
parameters.
- Removed "intendedmove" global (to game TU's unnamed namespace)
variable. It is now local to soloGameLoop()'s scope. Variable is now
named "player_intendedmove".
