+++ src/headers/global.hpp
@@ -4,7 +4,7 @@
 #include <string>
 
 using ull = unsigned long long;
-void getInput(char &);
+void getKeypressDownInput(char &);
 void wait_for_any_letter_input(std::istream &is);
 void newline(int n = 1);
 void clearScreen();
